Split release updates feed into components
Extract StatusUpdateCard, UpdateComposer, UpdatesList and types from
ReleaseStatusUpdatesFeed and move presentation + editor logic into
separate files. Optimize comment SSE refresh to only refetch opened
threads and avoid rendering an empty feed for non-managers.
